# internal/gen/model_error_response.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Маркетплейс API
 *
 * API для маркетплейса с полной бизнес-логикой: - Управление товарами (CRUD) - Управление заказами с валидацией - Промокоды и скидки - Аутентификация и авторизация (JWT) - Ролевая модель доступа (USER, SELLER, ADMIN) 
 *
 * API version: 1.0.0
 */

package gen


import (
	"time"
)



type ErrorResponse struct {

	ErrorCode ErrorCode `json:"error_code"`

	// Человекочитаемое описание ошибки
	Message string `json:"message"`

	// Дополнительные данные об ошибке
	Details map[string]string `json:"details,omitempty"`

	// Время возникновения ошибки
	Timestamp time.Time `json:"timestamp"`
}

// AssertErrorResponseRequired checks if the required fields are not zero-ed
func AssertErrorResponseRequired(obj ErrorResponse) error {
	elements := map[string]interface{}{
		"error_code": obj.ErrorCode,
		"message": obj.Message,
		"timestamp": obj.Timestamp,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertErrorResponseConstraints checks if the values respects the defined constraints
func AssertErrorResponseConstraints(obj ErrorResponse) error {
	return nil
}