# internal/gen/model_order_status.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Маркетплейс API
 *
 * API для маркетплейса с полной бизнес-логикой: - Управление товарами (CRUD) - Управление заказами с валидацией - Промокоды и скидки - Аутентификация и авторизация (JWT) - Ролевая модель доступа (USER, SELLER, ADMIN) 
 *
 * API version: 1.0.0
 */

package gen


import (
	"fmt"
)


// OrderStatus : Статус заказа
type OrderStatus string

// List of OrderStatus
const (
	ORDERSTATUS_CREATED OrderStatus = "CREATED"
	ORDERSTATUS_PAYMENT_PENDING OrderStatus = "PAYMENT_PENDING"
	ORDERSTATUS_PAID OrderStatus = "PAID"
	ORDERSTATUS_SHIPPED OrderStatus = "SHIPPED"
	ORDERSTATUS_COMPLETED OrderStatus = "COMPLETED"
	ORDERSTATUS_CANCELED OrderStatus = "CANCELED"
)

// AllowedOrderStatusEnumValues is all the allowed values of OrderStatus enum
var AllowedOrderStatusEnumValues = []OrderStatus{
	"CREATED",
	"PAYMENT_PENDING",
	"PAID",
	"SHIPPED",
	"COMPLETED",
	"CANCELED",
}

// validOrderStatusEnumValue provides a map of OrderStatuss for fast verification of use input
var validOrderStatusEnumValues = map[OrderStatus]struct{}{
	"CREATED": {},
	"PAYMENT_PENDING": {},
	"PAID": {},
	"SHIPPED": {},
	"COMPLETED": {},
	"CANCELED": {},
}

// IsValid return true if the value is valid for the enum, false otherwise
func (v OrderStatus) IsValid() bool {
	_, ok := validOrderStatusEnumValues[v]
	return ok
}

// NewOrderStatusFromValue returns a pointer to a valid OrderStatus
// for the value passed as argument, or an error if the value passed is not allowed by the enum
func NewOrderStatusFromValue(v string) (OrderStatus, error) {
	ev := OrderStatus(v)
	if ev.IsValid() {
		return ev, nil
	}

	return "", fmt.Errorf("invalid value '%v' for OrderStatus: valid values are %v", v, AllowedOrderStatusEnumValues)
}



// AssertOrderStatusRequired checks if the required fields are not zero-ed
func AssertOrderStatusRequired(obj OrderStatus) error {
	return nil
}

// AssertOrderStatusConstraints checks if the values respects the defined constraints
func AssertOrderStatusConstraints(obj OrderStatus) error {
	return nil
}