# pkg/get_t/api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Степан Фокин
  description: |
    API для маркетплейса с полной бизнес-логикой:
    - Управление товарами (CRUD)
    - Управление заказами с валидацией
    - Промокоды и скидки
    - Аутентификация и авторизация (JWT)
    - Ролевая модель доступа (USER, SELLER, ADMIN)
  title: Маркетплейс API
  version: 1.0.0
servers:
- description: Локальное окружение разработки
  url: http://localhost:8080/api/v1
- description: Продакшн
  url: https://api.marketplace.example.com/v1
...
paths:
...
  /orders:
    post:
      description: |
        Создаёт новый заказ с валидацией бизнес-правил:
        1. Ограничение частоты создания (последняя операция CREATE_ORDER)
        2. Проверка активных заказов (состояния CREATED, PAYMENT_PENDING)
        3. Проверка каталога (товары существуют и ACTIVE)
        4. Проверка остатков (stock >= quantity)
        5. Резервирование остатков
        6. Снапшот цен (price_at_order)
        7. Расчёт стоимости с учётом промокода
        8. Фиксация операции в user_operations

        - USER: разрешено
        - SELLER: запрещено (403)
        - ADMIN: разрешено
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
          description: Заказ успешно создан
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации данных
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Товар не найден
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Бизнес-ошибка (активный заказ, недостаточно товара, недопу\
            стимый статус)"
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка промокода
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Превышен лимит частоты создания заказа
      security:
      - bearerAuth: []
      summary: Создать новый заказ
      tags:
      - orders
  /orders/{id}:
    get:
      description: |
        Возвращает детальную информацию о заказе.
        - USER: только свои заказы
        - SELLER: запрещено (403)
        - ADMIN: любые заказы
      operationId: getOrderById
      parameters:
      - description: Уникальный идентификатор заказа
        example: 550e8400-e29b-41d4-a716-446655440001
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
          description: Заказ найден
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав (заказ принадлежит другому пользователю)
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Заказ не найден
      security:
      - bearerAuth: []
      summary: Получить заказ по ID
      tags:
      - orders
    put:
      description: |
        Обновляет позиции в заказе (только в состоянии CREATED).
        Выполняет те же проверки, что при создании:
        1. Проверка владельца
        2. Проверка состояния (только CREATED)
        3. Ограничение частоты обновления
        4. Возврат предыдущих остатков
        5. Проверка и резервирование новых позиций
        6. Пересчёт стоимости (включая промокод)
        7. Фиксация операции UPDATE_ORDER

        - USER: только свои заказы
        - SELLER: запрещено (403)
        - ADMIN: любые заказы
      operationId: updateOrder
      parameters:
      - description: Уникальный идентификатор заказа
        example: 550e8400-e29b-41d4-a716-446655440001
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
          description: Заказ успешно обновлён
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Ошибка валидации данных
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Необходима аутентификация
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Недостаточно прав
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Заказ не найден
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Бизнес-ошибка (недопустимый переход состояния)
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Превышен лимит частоты обновления заказа
      security:
      - bearerAuth: []
      summary: Обновить заказ
      tags:
      - orders
...
components:
  schemas:
...
    OrderStatus:
      description: Статус заказа
      enum:
      - CREATED
      - PAYMENT_PENDING
      - PAID
      - SHIPPED
      - COMPLETED
      - CANCELED
      type: string
...
    ErrorCode:
      description: Коды ошибок для контрактной обработки
      enum:
      - PRODUCT_NOT_FOUND
      - PRODUCT_INACTIVE
      - ORDER_NOT_FOUND
      - ORDER_LIMIT_EXCEEDED
      - ORDER_HAS_ACTIVE
      - INVALID_STATE_TRANSITION
      - INSUFFICIENT_STOCK
      - PROMO_CODE_INVALID
      - PROMO_CODE_MIN_AMOUNT
      - ORDER_OWNERSHIP_VIOLATION
      - VALIDATION_ERROR
      - TOKEN_EXPIRED
      - TOKEN_INVALID
      - REFRESH_TOKEN_INVALID
      - ACCESS_DENIED
      type: string
...
    OrderItemCreate:
      example:
        quantity: 2
        product_id: 550e8400-e29b-41d4-a716-446655440000
      properties:
        product_id:
          description: ID товара
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        quantity:
          description: Количество единиц товара
          example: 2
          maximum: 999
          minimum: 1
          type: integer
      required:
      - product_id
      - quantity
    OrderItemResponse:
      example:
        quantity: 2
        product_id: 550e8400-e29b-41d4-a716-446655440000
        price_at_order: 59990.0
        id: 660e8400-e29b-41d4-a716-446655440002
      properties:
        id:
          description: Уникальный идентификатор позиции
          example: 660e8400-e29b-41d4-a716-446655440002
          format: uuid
          readOnly: true
          type: string
        product_id:
          description: ID товара
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        quantity:
          description: Количество единиц товара
          example: 2
          type: integer
        price_at_order:
          description: Цена за единицу на момент заказа (снапшот)
          example: 59990.0
          format: double
          type: number
      required:
      - id
      - price_at_order
      - product_id
      - quantity
    OrderCreateRequest:
      example:
        promo_code: WINTER2024
        items:
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
      properties:
        items:
          description: Список позиций заказа
          items:
            $ref: "#/components/schemas/OrderItemCreate"
          maxItems: 50
          minItems: 1
          type: array
        promo_code:
          description: Промокод (опционально)
          example: WINTER2024
          pattern: "^[A-Z0-9_]{4,20}$"
          type: string
      required:
      - items
    OrderUpdateRequest:
      example:
        items:
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
      properties:
        items:
          description: Новый список позиций заказа
          items:
            $ref: "#/components/schemas/OrderItemCreate"
          maxItems: 50
          minItems: 1
          type: array
      required:
      - items
    OrderResponse:
      example:
        promo_code_id: c1b2c3d4-e5f6-7890-abcd-ef1234567892
        updated_at: 2024-03-04T15:30:00Z
        user_id: b1b2c3d4-e5f6-7890-abcd-ef1234567891
        total_amount: 119980.0
        discount_amount: 0.0
        created_at: 2024-03-04T15:30:00Z
        id: 550e8400-e29b-41d4-a716-446655440001
        items:
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
          price_at_order: 59990.0
          id: 660e8400-e29b-41d4-a716-446655440002
        - quantity: 2
          product_id: 550e8400-e29b-41d4-a716-446655440000
          price_at_order: 59990.0
          id: 660e8400-e29b-41d4-a716-446655440002
        status: CREATED
      properties:
        id:
          description: Уникальный идентификатор заказа
          example: 550e8400-e29b-41d4-a716-446655440001
          format: uuid
          readOnly: true
          type: string
        user_id:
          description: ID владельца заказа
          example: b1b2c3d4-e5f6-7890-abcd-ef1234567891
          format: uuid
          readOnly: true
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        promo_code_id:
          description: ID применённого промокода
          example: c1b2c3d4-e5f6-7890-abcd-ef1234567892
          format: uuid
          type: string
          nullable: true
        total_amount:
          description: Итоговая сумма заказа
          example: 119980.0
          format: double
          type: number
        discount_amount:
          description: Сумма скидки
          example: 0.0
          format: double
          type: number
        items:
          description: Позиции заказа
          items:
            $ref: "#/components/schemas/OrderItemResponse"
          type: array
        created_at:
          description: Время создания заказа
          example: 2024-03-04T15:30:00Z
          format: date-time
          readOnly: true
          type: string
        updated_at:
          description: Время последнего обновления заказа
          example: 2024-03-04T15:30:00Z
          format: date-time
          readOnly: true
          type: string
      required:
      - created_at
      - discount_amount
      - id
      - items
      - status
      - total_amount
      - updated_at
      - user_id
...
    ErrorResponse:
      example:
        error_code: PRODUCT_NOT_FOUND
        details:
          field: price
          value: "-100"
          constraint: must be > 0
        message: Product not found
        timestamp: 2024-03-04T14:30:00Z
      properties:
        error_code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          description: Человекочитаемое описание ошибки
          example: Product not found
          type: string
        details:
          additionalProperties:
            type: string
          description: Дополнительные данные об ошибке
          example:
            field: price
            value: "-100"
            constraint: must be > 0
        timestamp:
          description: Время возникновения ошибки
          example: 2024-03-04T14:30:00Z
          format: date-time
          type: string
      required:
      - error_code
      - message
      - timestamp
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: |-
        JWT access token для аутентификации.
        Формат: "Bearer {access_token}"

        Роли в токене (claim "role"):
        - USER: обычный покупатель
        - SELLER: продавец (может управлять товарами)
        - ADMIN: администратор (полный доступ)
      scheme: bearer
      type: http