# pkg/get_t/go/model_error_code.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Маркетплейс API
 *
 * API для маркетплейса с полной бизнес-логикой: - Управление товарами (CRUD) - Управление заказами с валидацией - Промокоды и скидки - Аутентификация и авторизация (JWT) - Ролевая модель доступа (USER, SELLER, ADMIN) 
 *
 * API version: 1.0.0
 */

package gen


import (
	"fmt"
)


// ErrorCode : Коды ошибок для контрактной обработки
type ErrorCode string

// List of ErrorCode
const (
	ERRORCODE_PRODUCT_NOT_FOUND ErrorCode = "PRODUCT_NOT_FOUND"
	ERRORCODE_PRODUCT_INACTIVE ErrorCode = "PRODUCT_INACTIVE"
	ERRORCODE_ORDER_NOT_FOUND ErrorCode = "ORDER_NOT_FOUND"
	ERRORCODE_ORDER_LIMIT_EXCEEDED ErrorCode = "ORDER_LIMIT_EXCEEDED"
	ERRORCODE_ORDER_HAS_ACTIVE ErrorCode = "ORDER_HAS_ACTIVE"
	ERRORCODE_INVALID_STATE_TRANSITION ErrorCode = "INVALID_STATE_TRANSITION"
	ERRORCODE_INSUFFICIENT_STOCK ErrorCode = "INSUFFICIENT_STOCK"
	ERRORCODE_PROMO_CODE_INVALID ErrorCode = "PROMO_CODE_INVALID"
	ERRORCODE_PROMO_CODE_MIN_AMOUNT ErrorCode = "PROMO_CODE_MIN_AMOUNT"
	ERRORCODE_ORDER_OWNERSHIP_VIOLATION ErrorCode = "ORDER_OWNERSHIP_VIOLATION"
	ERRORCODE_VALIDATION_ERROR ErrorCode = "VALIDATION_ERROR"
	ERRORCODE_TOKEN_EXPIRED ErrorCode = "TOKEN_EXPIRED"
	ERRORCODE_TOKEN_INVALID ErrorCode = "TOKEN_INVALID"
	ERRORCODE_REFRESH_TOKEN_INVALID ErrorCode = "REFRESH_TOKEN_INVALID"
	ERRORCODE_ACCESS_DENIED ErrorCode = "ACCESS_DENIED"
)

// AllowedErrorCodeEnumValues is all the allowed values of ErrorCode enum
var AllowedErrorCodeEnumValues = []ErrorCode{
	"PRODUCT_NOT_FOUND",
	"PRODUCT_INACTIVE",
	"ORDER_NOT_FOUND",
	"ORDER_LIMIT_EXCEEDED",
	"ORDER_HAS_ACTIVE",
	"INVALID_STATE_TRANSITION",
	"INSUFFICIENT_STOCK",
	"PROMO_CODE_INVALID",
	"PROMO_CODE_MIN_AMOUNT",
	"ORDER_OWNERSHIP_VIOLATION",
	"VALIDATION_ERROR",
	"TOKEN_EXPIRED",
	"TOKEN_INVALID",
	"REFRESH_TOKEN_INVALID",
	"ACCESS_DENIED",
}

// validErrorCodeEnumValue provides a map of ErrorCodes for fast verification of use input
var validErrorCodeEnumValues = map[ErrorCode]struct{}{
	"PRODUCT_NOT_FOUND": {},
	"PRODUCT_INACTIVE": {},
	"ORDER_NOT_FOUND": {},
	"ORDER_LIMIT_EXCEEDED": {},
	"ORDER_HAS_ACTIVE": {},
	"INVALID_STATE_TRANSITION": {},
	"INSUFFICIENT_STOCK": {},
	"PROMO_CODE_INVALID": {},
	"PROMO_CODE_MIN_AMOUNT": {},
	"ORDER_OWNERSHIP_VIOLATION": {},
	"VALIDATION_ERROR": {},
	"TOKEN_EXPIRED": {},
	"TOKEN_INVALID": {},
	"REFRESH_TOKEN_INVALID": {},
	"ACCESS_DENIED": {},
}

// IsValid return true if the value is valid for the enum, false otherwise
func (v ErrorCode) IsValid() bool {
	_, ok := validErrorCodeEnumValues[v]
	return ok
}

// NewErrorCodeFromValue returns a pointer to a valid ErrorCode
// for the value passed as argument, or an error if the value passed is not allowed by the enum
func NewErrorCodeFromValue(v string) (ErrorCode, error) {
	ev := ErrorCode(v)
	if ev.IsValid() {
		return ev, nil
	}

	return "", fmt.Errorf("invalid value '%v' for ErrorCode: valid values are %v", v, AllowedErrorCodeEnumValues)
}



// AssertErrorCodeRequired checks if the required fields are not zero-ed
func AssertErrorCodeRequired(obj ErrorCode) error {
	return nil
}

// AssertErrorCodeConstraints checks if the values respects the defined constraints
func AssertErrorCodeConstraints(obj ErrorCode) error {
	return nil
}